code divided in 3 parts and settings is done
