+++ static/admin_login.css
@@ -1,7 +1,7 @@
 
 body{
     background-color:rgb(223, 223, 223);
-    font-family: 'Times New Roman', Times, serif;
+    font-family:Arial, Helvetica, sans-serif;
 }
 
 ul {
