+++ static/view_list.css
@@ -1,6 +1,6 @@
 body {
   background-color:rgba(171, 172, 177, 0.493);
-  font-family: 'Times New Roman', Times, serif;
+  font-family:Arial, Helvetica, sans-serif;
 }
 ul {
   list-style-type: none;
