+++ templates/new_priority_adding.html
@@ -1,0 +1,63 @@
+
+{% block header %}
+<title>{% block title %}Priority Adding{% endblock %}</title>
+  <link rel="stylesheet" href="../static/home.css">
+  <style>
+    input[type=text], select {
+      width: 90%;
+      padding: 12px 20px;
+      margin: 8px 0;
+      display: block;
+      border: 1px solid #ccc;
+      border-radius: 4px;
+      box-sizing: border-box;
+    }
+    
+    input[type=submit] {
+      width: 45%;
+      background-color: #777977;
+      color: white;
+      padding: 14px 20px;
+      margin: 8px 0;
+      border: none;
+      border-radius: 4px;
+      cursor: pointer;
+    }
+    
+    input[type=submit]:hover {
+      background-color: #585c59;
+    }
+    
+    div {
+      width: 35%;
+      border-radius: 5px;
+      background-color: #f2f2f2;
+      padding: 20px;
+    }
+    </style>
+  {% endblock %}
+ {% extends 'main.html' %}
+  {% block content %}
+    <h1>New Priority Adding:</h1>
+    <div>
+        <form action="http://127.0.0.1:5000/new_priority_add" method="POST">
+          <label for="plabel">Priority Label:</label>
+          <input type="text" id="plabel" name="plabel" placeholder="Enter priority label">
+          <label for="plevel">Priority Level:</label>
+          <select id="plevel" name="plevel">
+            <option value="1">1</option>
+            <option value="2">2</option>
+            <option value="3">3</option>
+            <option value="4">4</option>
+            <option value="5">5</option>
+            <option value="6">6</option>
+            <option value="7">7</option>
+            <option value="8">8</option>
+            <option value="9">9</option>
+            <option value="10">10</option>
+          </select>
+        
+          <input type="submit" value="Submit">
+        </form>
+      </div>
+{% endblock %}
